Added a functional main menu scene (scene 0), with three buttons: PLAY (loads scene 1), OPTIONS (opens up options), and QUIT (quits application). Additionally, added a temporary scene for the game.
The main menu scene is called MenuScreen. Presumably, this is what the user first sees, so I set it to have a scene index of 0.

There are three buttons:

PLAY, which runs a script to move to scene index 1. Scene index 1 is where I assume we will be having the main game scene.

OPTIONS, which when pressed, moves to the options menu. The options menu currently only has a volume slider (which doesn't do anything right now) and a back to main menu button.

QUIT, which exits the application.

I added a "Game" scene which is the start of the actual game. It currently doesn't have anything other than a background and a button to go back to the main menu.